moodmate__frontend/requirements.txt
numpy==2.1.2
opencv_python_headless==4.10.0.84
pandas==2.2.3
Pillow<11.0.0
plotly==5.21.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.36.0
streamlit_extras==0.5.0